Refactor service file names for convention
