+++ frontend/src/components/UpdatePartner.vue
@@ -78,7 +78,7 @@
 <script lang="ts">
 import { EventBus } from "@/main";
 import VBpaButton from "@/components/BpaButton";
-import partnerService from "@/services/partnerService";
+import partnerService from "@/services/partner-service";
 import { Tag, UpdatePartnerRequest } from "@/services/partner-types";
 
 export default {
