+++ frontend/src/components/schema-add/ImportCredentialJsonLd.vue
@@ -105,7 +105,7 @@
 </template>
 <script lang="ts">
 import VBpaButton from "@/components/BpaButton";
-import adminService from "@/services/adminService";
+import adminService from "@/services/admin-service";
 import { EventBus } from "@/main";
 import { jsonLdService } from "@/services";
