+++ frontend/src/services/partner-service.ts
