+++ frontend/src/views/ProofTemplateView.vue
@@ -89,7 +89,7 @@
 <script lang="ts">
 import { EventBus } from "@/main";
 import AttributeGroup from "@/components/proof-templates/AttributeGroup.vue";
-import proofTemplateService from "@/services/proofTemplateService";
+import proofTemplateService from "@/services/proof-template-service";
 import VBpaButton from "@/components/BpaButton";
 
 export default {
